dj_backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.0.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
